Rename to fakeBaseQuery
